edited a lot made the transistions occure for play and menu button and continue and surrender timing is a bit strange
