--- main.py
@@ -1,42 +1,0 @@
-#!/usr/bin/python3
-"""Main - Clash of Conquerors game"""
-
-import pygame
-import sys
-
-from const import *
-from chess import Game
-from c_o_c import Chess
-"""Import the constants and the Game class"""
-
-
-class Main:
-    """Main class for game C.O.C."""
-    def __init__(self):
-        """initialize game and display the screen"""
-        pygame.init()
-        self.screen = pygame.display.set_mode((WIN_WIDTH, WIN_HEIGHT), pygame.FULLSCREEN)
-        pygame.display.set_caption("Clash Of Conquerors")
-        # Load and set the icon
-        icon_image = pygame.image.load('assets/icons/coc_icon_bigger.png')
-        pygame.display.set_icon(icon_image)
-        self.game = Game()
-
-    def mainloop(self):
-        """Main game loop"""
-        while True:
-            self.game.show_main_bg(self.screen)
-            self.game.show_main_menu(self.screen)
-            self.game.show_cursor(self.screen)
-
-            for event in pygame.event.get():
-                if event.type == pygame.QUIT:
-                    pygame.quit()
-                    sys.exit()
-
-            pygame.display.update()
-
-
-if __name__ == "__main__":
-    main = Main()
-    main.mainloop()
